+++ src/writer/json.rs
@@ -22,7 +22,7 @@ use crate::{
     cli, event,
     feature::ExpandExamplesError,
     parser,
-    writer::{self, basic::coerce_error},
+    writer::{self, basic::coerce_error, discard},
     Event, World, Writer, WriterExt as _,
 };
 
@@ -73,6 +73,28 @@ impl<Out: io::Write> Json<Out> {
     #[must_use]
     pub fn new<W: Debug + World>(output: Out) -> writer::Normalized<W, Self> {
         Self::raw(output).normalized()
+    }
+
+    /// Creates a new unnormalized [`Json`] [`Writer`] outputting [JSON][1] into
+    /// the given `output`, and suitable for feeding into [`tee()`].
+    ///
+    /// # Warning
+    ///
+    /// It may panic in runtime as won't be able to form [correct JSON][1] from
+    /// unordered [`Cucumber` events][2], until is [`normalized()`].
+    ///
+    /// So, either make it [`normalized()`] before feeding into [`tee()`], or
+    /// make the whole [`tee()`] pipeline [`normalized()`].
+    ///
+    /// [`normalized()`]: crate::WriterExt::normalized
+    /// [`tee()`]: crate::WriterExt::tee
+    /// [1]: https://github.com/cucumber/cucumber-json-schema
+    /// [2]: crate::event::Cucumber
+    #[must_use]
+    pub fn for_tee(output: Out) -> discard::Arbitrary<discard::Failure<Self>> {
+        Self::raw(output)
+            .discard_failure_writes()
+            .discard_arbitrary_writes()
     }
 
     /// Creates a new raw and unnormalized [`Json`] [`Writer`] outputting